pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling >= 1.13.0"]

[project]
authors = [
  {email = "danghuy1999@gmail.com", name = "Huy Nguyen"}
]
classifiers = [
  "Environment :: Web Environment",
  "Framework :: Django :: 4.0",
  "Framework :: Django :: 4.1",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.0",
  "Framework :: Django",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.9",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development"
]
dependencies = [
  "Django>=4",
  "django-admin-interface",
  "django-admin-sortable2",
  "django-filter",
  "django-jsonform",
  "django-localized-fields",
  "django-reversion",
  "django-solo",
  "djangorestframework>=3",
  "drf_spectacular",
  "martor",
  "psycopg"
]
description = "A simple django-based headless CMS."
dynamic = ["version"]
license = "MIT"
name = "django-headless-cms"
readme = "README.md"
requires-python = ">=3.9"

[project.urls]
Documentation = "https://django-headless-cms.readthedocs.io/"
Homepage = "https://github.com/huynguyengl99/django-headless-cms"
Repository = "https://github.com/huynguyengl99/django-headless-cms"

[tool.black]
exclude = '''
/(
    \.git
 |  \.pytest_cache
 |  \.vscode
 |  __pycache__
 |  .venv
 |  build
 |  coverage
)/
'''
line-length = 88
preview = true

[tool.coverage.run]
omit = [
  "headless_cms/utils/martor_custom_upload.py",
  "tests/*"
]

[tool.hatch.build]
exclude = [
  "/.*",
  "/scripts",
  "/tests"
]

[tool.hatch.build.targets.wheel]
packages = ["headless_cms"]

[tool.hatch.version]
path = "headless_cms/__init__.py"

[tool.ruff]
src = ["headless_cms", 'tests']

[tool.ruff.lint]
ignore = [
  "E501"  # line too long, handled by black
]
select = [
  "B",  # flake8-bugbear
  "C",  # flake8-comprehensions
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "N",  # pep8-naming
  "PL",  # pylint
  "Q",  # flake8-quotes
  "UP",  # pyupgrade
  "W"  # pycodestyle warnings
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR0913", "PLR2004"]
"tests/test_utils/base.py" = ["N802"]

[tool.tomlsort]
all = true
in_place = true
spaces_before_inline_comment = 2
